+++ Tests/PropTests/ItemBaseTests.cs
@@ -1,9 +1,9 @@
 using NUnit.Framework;
 using WForest.Exceptions;
 using WForest.Factories;
-using WForest.UI.Props.Grid.ItemProps;
-using WForest.UI.Widgets.Interfaces;
+using WForest.Props.Grid.ItemProps;
 using WForest.Utilities;
+using WForest.Widgets.Interfaces;
 using static Tests.Utils.HelperMethods;
 
 namespace Tests.PropTests