+++ Tests/PropTests/TextPropsTests/FontSizeTests.cs
@@ -5,9 +5,9 @@ using Moq;
 using NUnit.Framework;
 using WForest.Exceptions;
 using WForest.Factories;
-using WForest.UI.Props.Text;
-using WForest.UI.Widgets.BuiltIn;
+using WForest.Props.Text;
 using WForest.Utilities.Text;
+using WForest.Widgets.BuiltIn;
 
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]