+++ Tests/WidgetTests.cs
@@ -1,11 +1,10 @@
 using Moq;
 using NUnit.Framework;
 using WForest.Factories;
-using WForest.UI.Props.Interfaces;
-using WForest.UI.Widgets;
-using WForest.UI.Widgets.BuiltIn;
-using WForest.UI.Widgets.Interfaces;
+using WForest.Props.Interfaces;
 using WForest.Utilities;
+using WForest.Widgets;
+using WForest.Widgets.Interfaces;
 
 namespace Tests
 {