+++ WForest/src/Props/Dragging/FixX.cs
@@ -1,0 +1,11 @@
+using WForest.Props.Interfaces;
+
+namespace WForest.Props.Dragging
+{
+    /// <summary>
+    /// Property to make the widget unmovable on the X axis.
+    /// </summary>
+    public class FixX : IProp
+    {
+    }
+}