+++ WForest/src/Props/Interfaces/IProp.cs
@@ -1,0 +1,9 @@
+namespace WForest.Props.Interfaces
+{
+    /// <summary>
+    /// Base interface for properties of widgets.
+    /// </summary>
+    public interface IProp
+    {
+    }
+}