+++ WForest/src/Utilities/Text/FontStore.cs
@@ -17,7 +17,6 @@ namespace WForest.Utilities.Text
         /// Get or set the DefaultFont. Set the default font before using the FontStore.
         /// It will be used whenever text is used and a custom font is not required.
         /// </summary>
-        /// <exception cref="FontStoreNotInitializedException"></exception>
         public static Font DefaultFont
         {
             get
@@ -68,7 +67,7 @@ namespace WForest.Utilities.Text
         private static void CheckIfInit()
         {
             if (_defaultFont == null)
-                throw new FontStoreNotInitializedException(
+                throw new FontNotFoundException(
                     "No default font found, assign one first to FontStore.DefaultFont");
         }
     }